% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMC.R
\name{SAMC}
\alias{SAMC}
\title{Stochastic Approximation Monte Carlo (SAMC) Sampler}
\usage{
SAMC(nv, energy, domain = c(-Inf, Inf), partition = seq(-100, 100,
  length.out = 9), vecpi = rep(1/10, 10), tau = 1, niter = 20000,
  t0 = 200, xi = 2/3, stepsize = 1, trange = c(-Inf, Inf))
}
\description{
Stochastic Approximation Monte Carlo (SAMC) Sampler
}
\examples{
\dontrun{
##### Two-Dimensional Multimodal sampling
## Define negative log-density function as R function
func_r = function(x){
x1 = x[1]; x2 = x[2];
val1 = (-(x1*sin(20*x2)+x2*sin(20*x1))^2)*cosh(sin(10*x1)*x1);
val2 = (-(x1*cos(10*x2)-x2*sin(10*x1))^2)*cosh(cos(20*x2)*x2);
return(val1+val2);
}

## Other Experimental Setting
ex_part   = c(-Inf,seq(from=-8,to=0,by=0.2))
m         = length(ex_part)-1
ex_temp   = 1.0
ex_step   = (0.25)
ex_vecpi  = as.vector(array(1/m,c(1,m)))
ex_domain = c(-1.1,1.1)
ex_niter  = 200000

## Run The Code
res = SAMC(2,func_r,partition=ex_part,tau=ex_temp,stepsize=ex_step,vecpi=ex_vecpi,domain=ex_domain,niter=ex_niter)

## Visualize
select = seq(from=101,to=ex_niter,by=100) # 100 burn-in, 1/100 thinning 
par(mfrow=c(1,2))
plot(res$samples[select,1], res$samples[select,2],xlab='x',ylab='y',main='samples')
barplot(as.vector(res$frequency/sum(res$frequency)),
        main="visiting frequency by energy partition",
        names.arg=ex_part[2:(m+1)], xlab="energy")
}

}
\author{
Kisung You
}
